Split agent-history-tab.tsx (1164 lines) into three focused files
Extract the detail view and timeline components into separate files:
- agent-history-detail.tsx (454 lines): AgentHistoryDetail, DetailTabs, DurationBar
- agent-history-timeline.tsx (319 lines): EventTimeline, FeedbackTimeline, FeedbackItemRow, shared constants
- agent-history-tab.tsx (424 lines): AgentHistoryList + main export (routing shell)
